// protobuf/src/common_pb.ts
// @generated by protoc-gen-es v2.6.2 with parameter "target=ts"
// @generated from file common.proto (package system, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file common.proto.
 */
export const file_common: GenFile = /*@__PURE__*/
  fileDesc("Cgxjb21tb24ucHJvdG8SBnN5c3RlbSIwCghSb29tVXNlchISCgpzZXNzaW9uX2lkGAEgASgDEhAKCHVzZXJuYW1lGAIgASgJYgZwcm90bzM");

/**
 * @generated from message system.RoomUser
 */
export type RoomUser = Message<"system.RoomUser"> & {
  /**
   * The user's session ID.
   *
   * @generated from field: int64 session_id = 1;
   */
  sessionId: bigint;

  /**
   * The user's username.
   *
   * @generated from field: string username = 2;
   */
  username: string;
};

/**
 * Describes the message system.RoomUser.
 * Use `create(RoomUserSchema)` to create a new message.
 */
export const RoomUserSchema: GenMessage<RoomUser> = /*@__PURE__*/
  messageDesc(file_common, 0);